pyproject.toml
[project]
name = "deckard"
version = "0.1.0"
description = ""
authors = [
    {name = "Bill Tyros",email = "me@strategineer.com"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "discord.py",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
